refactor: set for debug only tests
